Moving file to middleware folder
